fix: build 오류 해결
