+++ app/components/diary-book/index.ts
@@ -1,3 +1,4 @@
+export { default as DiaryBookCreateCoverForm } from './create-cover-form';
 export { default as DiaryBookCreateCoverPalette } from './create-cover-palette';
 export { default as DiaryBookCreateSettingPeriod } from './create-setting-period';
 export { default as DiaryBookCreateSettingTitle } from './create-setting-title';
